ai-service/requirements.txt
# Core Server and API Frameworks
fastapi
uvicorn
pydantic

# Data Processing and Numerical Operations
pandas
numpy

# LLM Connections
requests

# Visualization
plotly

# Machine Learning
tensorflow
joblib

# Database
psycopg2-binary
python-dotenv

# File handling
python-multipart

# For CORS
starlette

# Optional: For audio/STT (can be installed later)
# faster-whisper
# ffmpeg-python
